requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
croniter==2.0.1
requests==2.31.0
python-dotenv==1.0.0